Move Terraform config to its own repo.
The TF config will live in its own private repo in an Azure DevOps
workspace for now. While it doesn't contain secrets, it does contain
plenty of information we wouldn't like to advertise, like IP addresses
and CIDR ranges. Once the TF config is further evolved, we can pull
the modules back into this repo. For now, it will all live in one place
for ease of use.