K8s configuration for TF-managed infrastructure.
This adds overlays for the sites running in the new TF-managed
infrastructure. The sites are namespaces as "staging" and "master" and
will correspond to the branches with those names. This also updates the
CircleCI config with values for the new container registry and
Kubernetes cluster.

One approach I've taken that's new is that the master overlay is based
on the staging overlay. This should cut down on the amount of
boilerplate copied and pasted around.

This doesn't remove the config for the old sites just yet. I think that
should happen separately when we've confirmed that the entire migration
is complete and working as expected.