Copy the directory name correctly in the Docker build.
The top-level module should be named "atat" inside the Docker container.
Confusingly, the application code is inside "/opt/atat/atst", with the
top-level application module at "/opt/atat/atst/atat". That directory
path is a holdover from when this application was a series of
microservices. Eventually we should truncate it to something sensible
and descriptive.