Add Redis hostname back for pwdev k8s environment.
The DNS resolution issue for the Kubernetes cluster appears to be
resolved, and so we can add the Redis hostname back to the config. I'm
only committing this single hostname change at the moment because I did
it while debugging Redis downtime.

The DNS issue was resolved by associating a route table that AKS
provisions for itself (that resolves the cluster's IP space to node IPs)
to the subnet the AKS cluster resides in. We'll follow up on managing
this in the Terraform configuration separately.

In the future we should:
- Restore the Postgres hostname
- Configure Redis and Postgres TLS connections to do full verification
  against their respective hostnames

PT: https://www.pivotaltracker.com/story/show/171469501