Add additional config for Hybrid interface to master.
The Hybrid interface requires additional configuration to run. This is
set in the form of ConfigMap values and FlexVol secrets. There are some
pieces of data--notably IDs--which isn't secret per se but is best not
committed to an open source repo, so I decided to make those FlexVol
secrets as well.