The Hybrid CSP interface should not rotate the admin creds.
In a production CSP setup, each portfolio would have a separate
administrator account that ATAT would use for some initial
bootstrapping. When ATAT is done with portfolio setup, it needs to reset
the credentials for that admin account. In the Hybrid CSP setup, there
is only one administrator account for every portfolio. Because this is
the case, ATAT should not reset the admin creds when using the Hybrid.
When it does, it locks itself out of the account and prevents any
further provisioning.