Script for removing Hybrid app registrations.
This script will delete app registrations added by the
HybridCloudInterface. It could be extended to delete additional entities
that provider can provision, like billing users and management groups. I
opted to handle app registrations to start because those entities are
given a huge number of permissions and it seems unwise to have too many
in a tenant.