Update query for finding CLINs to resend to the CSP.
The query method for finding CLINs that need to be sent to the CSP needs
to check whether a portfolio is done being provisioned (i.e., its state
machine has the final state of COMPLETED). There are two cases where we
need to send CLIN information to the CSP:

- during the initial provisioning of the portfolio.
- when CLINs are added or edited after a portfolio has been provisioned.

The query and Celery task it's used by exist to handle the second case.
Before this update, ATAT would try to resend CLINs for portfolios that
had not finished provisioning and were missing necessary information
(such as the billing profile).

This commit also updates the dictionary access method to the portfolio's
csp_data in the Celery task, since we do not handle cases where we don't
find the relevant data.