Include filename in downloaded task order dictionary.
When ATAT downloads a task order, it should add the the filename to the
dictionary of data it returns for the Azure Blob object. This is useful
for including in email attachments. The "name" property is the name ATAT
assigned the object in Azure blob storage, which is not the same as the
filename (the object name must be unique within the storage container,
so we use a GUID). ATAT's frontend should add the filename as metadata
to the storage object and it should still be available to ATAT when
downloading the task order. We'll supply a default filename for cases
where the filename is not available.

This also updates:
- the mock Files interface so that it provides a filename
- the tests for `send_task_order_files`, the Celery task responsible for
  downloading and transmitting task orders to the CSP