Fix CRL check test.
DISA seems to have deprecated or removed the DMDNSIGNINGCA_1 CA/CRL. The
CRL path for it now returns a 404. This was causing our nightly CRL
CI check to fail. I removed the CRL from our list. I also updated the
error logging for the CRL downloads to include the class of the error
that was thrown.