Fix bug in EnvironmentRoles.disable
Previously, this method would only invoke the cloud interface method for
disabling an Azure role assignment if the associated environment did
_not_ have an associated management group in Azure (signified by a value
in the `cloud_id` column for that environment). This doesn't make any
sense and is a mistake that crept in at some point.

This commit also introduces a property to the portfolio model to allow
easier access to the tenant ID for that portfolio. This way callers
don't have to know the implementation details for that data.