Dry Azure Testing -- mock resp and creds
Even though we had a helper function to make mock responses, we were
sometimes building them by hand in tests. Now we only use the helper.

It was also common to mock Adal's acquire_token_with_client_credentials
method to return a mock token. Now it's defined at the Mock level so we
don't need to remember to Mock it.