Move mocking and patching to mock_azure fixture
Many unit tests would mock or patch the same functions on the Azure
cloud provider -- mostly related to acquiring access tokens. This commit
moves as much as that patching as possible to the mock AzCP fixture.
Now, we only Mock or patch to simulate failures, which reduces some
noise in the tests.