Remove default section for make_config's param
If make_config is supplied a `direct_config` param, it uses read_dict
to directly write those config options.

Previously, the call to read_dict wrote to the "default" section of the
config by default. This commit removes that option so that one can write
config options to any section.