Add hidden input for revoking env access form
Revoking environment access uses a form with checkbox input. When an
environment was revoked, the "checkbox" would be checked, while the
other environments' checkboxes were left unchecked. This means that when
the form was POSTed to the app, only the revoked, i.e. checked
envionments were included in the form data, since unchecked checkboxes
aren't included in form data by default. This mismatch caused issues when
we went to build the form with the POSTed form data and existing environment
data.

To overcome this, when a checkbox is unchecked, we should include a
hidden input with a falsy value for the form field. This way, the number
of environments and input values for those environments will be equal,
and WTForms will be able to process that input accurately.