Update uWSGI logging configuration.
This change follows advice found here:

https://www.velebit.ai/blog/tech-blog-uwsgi-json-logging/

Any logs from the ATAT application are passed through from uWSGI
untouched. This fixes the issue where ATAT logs nested inside uWSGI logs
were output as escaped JSON strings, rather than JSON. The issue is
discussed in detail on this previous PR:

https://github.com/dod-ccpo/atst/pull/740

This commit moves the uwsgi.ini file to the deploy/azure directory and
uses a [kustomize
configMapGenerator](https://kubernetes.io/docs/tasks/manage-kubernetes-objects/kustomization/#configmapgenerator)
to create the ConfigMap. This way the config can be a regular INI file.
I've also symlinked it into the repo root so that it's available for the
Docker build. We may want to use a configMapGenerator for the NGINX
config in the future.

Notes:

- The linked blog post recommends using an additional, custom uWSGI
  plugin to escape the host and uri values, which are two fields that
  may not be JSON-escaped. I tried doing this but our Alpine base image
  does not have gcc available for compiling C code, so commands like
  `uwsgi --build-plugin escape_json.c` fail. I did not think it was
  worth the effort to resolve that issue right now.
- The above point does mean that our use of the host field in uWSGI logs
  may break the JSON encoding, but the risk seems minimal. I would
  rather solve the problem when it appears then oversolve now.
- This commit also rearranges the uWSGI logs slightly so that "source"
  appears as the first property.
- I've updated the environment variables required to run the ATAT
  container in the docker-compose.yml file.