Add dependabot config.
Github's dependabot integration can help with upgrading dependencies. It
examines known dependency files and determines which, if any,
dependencies can be upgraded. It opens pull requests for updates.

This configuration specifies that dependabot should open PRs for both
Python and Javascript. It also allows dependabot to auto-merge patch
releases for dependencies that use semantic versioning. If this is too
aggressive we can adjust the settings.