Add missing hybrid marks
Also changes one skip to xfail. Pytest comments: "A skip means that you
expect your test to pass only if some conditions are met, otherwise
pytest should skip running the test altogether. ...  A xfail means that
you expect a test to fail for some reason. A common example is a test
for a feature not yet implemented, or a bug not yet fixed." The latter
seems more appropriate for this case.