Refine Hybrid implementation to work with new Key Vault.
The Terraform code was recently updated to include the Key Vault and app
registration used for managing customer AAD tenants. I ran our Hybrid
test suite using the URL for the new KV and creds for the new app
registration and encountered a few errors that are fixed here.

There were two app registrations used for running the Hybrid test suite:
1. the one that has access to the Key Vault where tenant credentials are
   stored
2. one that lives in the Hybrid tenant and exists strictly for testing
   purposes
I've added a third that has the "Invoice Section Reader" role for our
invoice section. We can't give that permission to 1 since 1 is managed
by TF, and giving that perm to 2 would require making 2 a multi-tenant
app. Instead, I've added a new app registration and accompanying config
for the purposes of running the `test_get_reporting_data` integration
test.