Add sleep to Hybrid portfolio provisioning test.
There appears to be a race condition in the
test_hybrid_provision_portfolio test. The test flickers on successive
runs. When it fails, the new app registration cannot get a token in the
create_initial_mgmt_group method. The step before
create_initial_mgmt_group (create_principal_admin_role) assigns the
Global Admin role to the new app registration. It's possible that the
role assignment doesn't propagate instantly and so sometimes a subsquent
authorization call will fail. This is just speculation, though, and we
should research the issue.