Clean up unused k8s pod identity solution.
In a previous iteration, our k8s config relied on a pod identity
strategy for authenticating to Azure Key Vault in order to populate
application configuration secrets via the FlexVol plugin. We have since
migrated to a different identity strategy with FlexVol which relies on
the managed identity associated with the virtual machine scale set
underlying the AKS cluster. This commit removes references to unused
config and updates the documentation.