Use printf style formatting for log messages
We were relying on modern string formatting methods like f-strings
and .format() to interpolate values in our log messages. The logging
module suggests to use printf style formatting though. Now, we pass a
templated message for the message arg and use positional arguments as
values to be interpolated. This way, "formatting of message arguments is
deferred until it cannot be avoided."

https://docs.python.org/3/howto/logging-cookbook.html#formatting-styles
https://docs.python.org/3/howto/logging.html#optimization