Enforce nesting of application management groups.
When we create management group, we transmit information about what
management group should be its parent. For an application management
group, its parent should be the portfolio management group. These parameters
in the request body do not seem to be respected by Azure at the moment.
It is creating application management groups as children of the tenant's
root management group, which is not the hierarchy we want for our Hybrid
implementation.

Making a subsequent PATCH request on the application management group to
update its parent fixes the issue, but we should not have to do this. I
think that we should use this as a temporary fix and investigate why the
Azure resource management API doesn't seem to respect the initial set of
parameters we give it.

This commit introduces a new test class for the hybrid suite.
TestIntegration. The tests are highly dependent on their order here.
The first test provisions a portfolio in Azure. The second test for
application creation makes an application as a child of that portfolio.
It uses the app registration associated with the portfolio to create the
application's management group. It also adds new code to check that the
application management group is nested under the correct parent.

Notes

- The TestIntegration class implements its own version of the database
  session fixture. Since database entities implemented as fixtures need
  to be shared between tests in the class, the database session needs to
  be scoped to the test run session.