Fix environment management group nesting.
This updates the value we save for the environment cloud_id to the fully
qualified ID for the Azure management group, which is the change
previously applied to the application cloud_id in 60aef421e7. This also
moves the hybrid test for environment creation into the TestIntegration
class and ensures that the environment is created as a child of the
application management group.