Pass all upload config to frontend.
This updates both the code and build process so that it's no longer
necessary to give details about the task order upload location to the
frontend when the JS bundle is compiled. Now, all the config the JS
needs to upload a file is passed to it by the task_orders.upload_token
endpoint. This removes some complexity from the build process and makes
the build less brittle.

Encompasses the following changes:

- Change buildUpload function so that it takes cloud provider and
  uploader class config as arguments.
- Update upload_token endpoint an JS to send CSP and uploader config
- Remove script and commands for pre-populating upload config values
  (script/write_dotenv, the accompanying ARGs in the Dockerfile, and the
  build-args passed to the Docker build in CI are no longer necessary)