Fix cluster_migration script for sh.
script/cluster_migration has to run with /bin/sh for CI and is currently
failing. I was using an array to check all of the variables that need to
be defined. It seems that sh doesn't natively support arrays and it
requires some fancy footwork to do anything similar with strings. Rather
than complicate the script or burn too much time on it, I've decided to
just use a conditional test for each of the variables we expect to be
defined.