Remove _creation & verify methods for subscription
Instead of using the _creation, _verificaiton pattern for creating
subscriptions, we just use a method create_subscription to handle
subscription creation. Though this is a long running job, this method
just ensures that the initial call to kick off the long running job is
successful. It does not monitor the status of the job.

Some logic to track this long-running job status may be put back when we
create a default subscription in an environment.