Simplify K8s port usage.
Turns out the ports that our NGINX container listens on for its 4 server
blocks do not need to be configurable across namespaces. Previously, we
thought that re-using the same port in two separate namespaces would
lead to port clashes, but it turns out this is not the case. K8s uses
namespace as part of its internal network resolution. This commit
removes all configuration for those ports.

Notes:
- This also updates script/k8s_config so that PORT_PREFIX is not a
  required environment variable.
- I've removed a couple old, hard-coded IPs for the load balancers from
  a previous iteration of the cluster.