Remove seprate pendulum install and poetry env use
`poetry env`: Removing this was the fix that allowed the container to
build. We set python3 to python 3.7.3 in the base image for this
dockerfile, so we don't need to tell poetry to use a specific version.

`pendulum`: This was an issue specific to Alpine builds, which was our
old base image. This step doesn't seem to be necessary anymore not that
we've switched over to a RHEL-based image.