Replace user token util w/ generic principal util
We previously moved get_user_principal_token_for_scope from the Azure
Cloud Provider class to a utils class so that it could be resued outside
the context of the Azure Cloud Provider.

In doing work for [AT-5312], I realized that it would be helpful to
obtain the token for a service principal outside of the Azure Cloud
provider in some cases as well. Additionally, I noticed that the bodies
for both _get_service_principal_token and
_get_user_principal_token_for_scope were very similar in that they both
sent a request to the same URL, just with a different payload.

In this commit, I reorganized these functions so that both the service
principal and user principal token functions call a common function that
handles POSTing a payload to the property endpoint. This common function
also replaces the specific user principal token function. It also
introduces two new cloud model classes that represent the slightly
different payloads for a service principal token vs a user principal
token.