Use Location header to get completed mgmt group
Creating management groups is an async operation. After sending
the initial request to create a management group, Azure returns a
response that contains two important headers: Azure-AsyncOperation and
Location. The former reports the status of the async job, and until now
was the only header we needed to track.

GETing the URL specified in the Azure-AsyncOperation header would return
a payload like:

```
{
    "id": "/providers/Microsoft.Management/operationResults/create/asyncOperation/status/managementGroups/<uuid>",
    "name": "<uuid>",
    "status": "Succeeded",
    "properties": {
        "tenantId": "<tenant_id>",
        "displayName": "Hybrid :: Application Name 0",
        "details": {
            "version": 1,
            "updatedTime": "2020-07-31T20:32:29.9423339Z",
            "updatedBy": "<user_uuid>",
        },
    },
}
```

Things to note here:
 - the id contains "operationResults/create/asyncOperation/status",
   which we don't want
 - there is no information about the parent, which we expect in the
   response

Even though the docs say "If your operation returns this value, always
use it (instead of Location) to track the status of the operation", I
found that by GETing the url provided by the Location header, we receive
the expected payload:

```
{
    "id": "/providers/Microsoft.Management/managementGroups/<uuid>",
    "type": "Microsoft.Management/managementGroups",
    "name": "<uuid>",
    "properties": {
        "tenantId": "<tenant_id>",
        "displayName": "Hybrid :: Application Name 0",
        "details": {
            "version": 1,
            "updatedTime": "2020-07-31T20:32:29.9423339Z",
            "updatedBy": "<user_uuid>",
            "parent": {
                "id": "/providers/Microsoft.Management/managementGroups/<tenant_id>",
                "name": "<tenant_id>",
                "displayName": "Tenant Root Group",
            },
        },
    },
}
```

I _believe_ that just relying on the Azure-AsyncOperation header url was
enough up to this point, but something must have changed on Azure's end
such that this url no longer provided the same payload.

Now, we use the Azure-AsyncOperation header URL to track the status of
the job. When we get a status of "Succeeded", we then use the Location
url to get the full payload for the created management group.