Disable interpolation for config parsing
By default, python's ConfigParser will try to resolve % as a token that proceeds a reference in the config file. Since many of our config values are generated (.e.g passwords) they may contain % and can cause issues deploying the code. This disables the interpolation entirely so we don't have to try to conform our values to the default parser implementation.