Add triggers to account for incomplete async ops
Some API calls in Azure kick off "async operations". In the initial call
to and endpoint, we receive a response with headers that contain a URL
to check on the status of the ongoing operation.

We handle portfolio provisioning stages that require async operations by
using _CREATION and _VERIFICATION steps in the state machine. The
_CREATION step handles the initial call to the endpoint and saves the
URL to check, and the _VERIFICATION step hits the this URL for the
payload of the operation.

We made an assumption that by the time we got to the _VERIFICATION step,
the async operation would be complete and we'd be able to return the
desired payload. But, there's a chance that we could try a the
verification URL and the job isn't completed yet. This would mean that
we'd either falsly report that the job was successful before it
completed, or possibly that the payload would not have all the expected
values and a Pydantic validation error would be raised.

In this commit, we add infrastructure to the state machine to account
for this case of incomplete operations. If we try a URL in the
verification step and the operation is still "In Progress", we reset the
state machine from the current stage's _IN_PROGRESS state to the
previous stage's _CREATED state. This effectively allows for the
verification URL to be tried again after the next query for portfolios
pending provisioning.