Full verification of SSL/TLS connections to Postgres.
The strictest PostgreSQL connection option (PGSSLMODE) available is
"verify-full".  This forces the client to verify the certificate the
server presents against a known certificate authority. The CA is
specified with PGSSLROOTCERT.

https://www.postgresql.org/docs/10/libpq-connect.html#LIBPQ-CONNECT-SSLMODE

Because of DNS issues we were experiencing within the
cluster at one point, the client could not resolve the full hostname of
the Postgres server every time. We resorted to using the IP address of
the server and not using PGSSLMODE=verify-full. That DNS issues is
resolved. This commit reverts us to using the default value set in the
Kubernetes config, PGSSLMODE=verify-full.