Full verification of TLS connection to Redis server.
This change allows us to configure a full verification of the Redis
SSL/TLS connection. Under the hood, the Python client library for Redis,
redis-py, is using the Python ssl module to make the connection. We can
pass the same configuration values we would normally provide to the ssl
module. In ATAT's config, these are represented by REDIS_SSLMODE and
REDIS_SSLCHECKHOSTNAME. The former will force the client to verify the
server's certificate against trusted certificate authorities, while the
latter verifies that the server's hostname if what's listed on the
server's certificate. These are represented as query arguments on
the URI we construct for the Redis connection.

Unfortunately, Celery requires a Redis URI for the BROKER_URL setting,
so we can't use a simple dictionary of options to make the connection.
We have to do some ugly string formatting instead.

I've updated the documentation and Kubernetes config, too.