Move AAD tenant hostname into module constant.
This value was being kept as Flask application config, but we don't
actually change it. Since we use it in the context of a utility
function, it would be preferable if we didn't need to make an instance
of the Flask app with config in order to access that value.