Add uWSGI back into container for integration tests.
Our integration tests use the first stage of the ATAT docker build
(usually called "builder"). Previously, this stage included uWSGI. My
previous PR removed uWSGI from that stage since it's not strictly
necessary for the build process. This inadvertantly broke the
integration tests. Rather than put it back in the docker file, I've made
it so that the integration test command installs it in the builder
image. That way the integration tests are responsible for adding their
own dependencies, rather than there being an unspoken dependency on it
being pre-installed in that stage.