Eagerly load PDF data in send task orders query
When calling TaskOrders.get_for_send_task_order_files in the
send_task_order_files celery job, we were getting errors because we were
unable to laziliy load pdf data through task order objects in the
context of a celery worker. Now, we eagerly load that data so that PDF
data can be used when needed in the task.