Update Celery loglevel argument.
Celery no longer accepts lowercase values for the command line
`--loglevel` flag. This commit update the K8s config and local script to
supply the uppercase value.

Note that the Celery command exits with an error if you supply the wrong
value, so Celery was not running in the K8s cluster.