Remove requirements.txt from root
In 3b75b6fc2d2e9aa5f1cba54e45532fc74ed42798, we moved a requirements.txt
file from the project root to the ops directory because these
requirements are meant to specify ops dependencies only. When we made
this move, we forgot to delete the original requirements.txt from the
root. This commit does that.