Make small optimizations to application Dockerfile
This commit does not represent a comprehensive Docker optimization
effort. Rather it grabs low hanging fruit to make the build of the
application container slightly faster:
1. Like in 8df2cf6dd5840a4534f87cc78366e075969a7383, we reduce the number
of layers by combining RUN steps where possible
2. One place where we actually split RUN and COPY steps apart involves
   installing project dependencies. Now, we selectively COPY Python and
Javascript dependency lockfiles, install the dependencies, and create a
production JS bundle before COPYing the rest of the source. This way, we
limit invalidating the cached dependency layers to when we acutally
change dependencies.

There are probably some dependencies still present in the application
Dockerfile that are uncecessary or could be moved to the base image. The
issue is that we install some dependenices here, but then install
other tools directly in CI/CD pipelines. Since the story for this ticket
focuses specifically on our Dockerfiles, and since we don't know what
our CI/CD setup will look like for production, we held off on making
changes that would reqiure modifying our Circle CI config. There are
still plenty of gains to be made to optimize our Docker build process
and pipelines, but they'll come later when some of these questions are
answered.