Ensure local certs created where expected
If the SSL directory did not exist, it was not created and the cd
failed. This resulted in the certificates being placed in the current
directory instead. The secure_server script loads the certificates
specifically from the ssl/ folder. That directory needs to be created
and the certificates need to be stored there. There is not a particular
need to cd; this can be done instead with just specifying that directory
in the openssl arguments.